New option: delay (milliseconds)
Adds a new option "delay" which hides the animation for the specified amount of milliseconds.
This is useful in the case where the action may happen quickly, causing a flash as the spinner quickly appears and disappears. With delay you can give the user the "please wait" feedback after a certain delay has passed.